+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - jrl_release: add `--check-version --check-tag vX.Y.Z` by @arntanguy
 - jrl_release: add `DebianChangelogVersionExtractor` by @arntanguy
 - jrl_check_python_module: fall back to `importlib.metadata` when the module has no `__version__` by @ahoarau
+- jrl: change jrl_configure_default_install_dirs() to macro to make GNUInstallDirs in caller's scope by @ahoarau
 
 ## [2.1.0] - 2026-07-03